Build a populationExpressionObj from a GatingSet

Build a populationExpressionObj from a GatingSet

PEOFromGatingSet(gs, annotation = NULL, populations = NULL,
  samplePop = 4000, objId = NULL, mapVar = NULL)

Arguments

gs
- a gatingSet

annotation
- Annotations for each sample, where one column = sampleNames(gs). If NULL, then it will attempt to grab annotation from the phenoData slot of the gatingSet.

populations
- A list of populations (must correspond to populationNames in gs). If NULL, will just set populations with all populations in gatingSet

samplePop
- Number of cells per population to sample. If NULL, returns all cells in population.

objId
- A unique object identifier used to avoid namespace collisions. If NULL, a unique ID will be generated for the populationExpressionObj

Value

populationExpressionObj