SuperClass for other DataObj Classes

SuperClass for other DataObj Classes

commonDataObj

Format

R6Class object.

Value

Object of R6Class with methods for getting/setting features for flowDashboard.

Fields

annotation

Annotation (can be extracted as phenoData from a GatingSet) as data.table.

subsetOptions

set which columns in annotation to use for subsetting. Set by setSubsetAndSortOptions().

subsetOptionList

named list, where every entry corresponds to levels in a column in annotation.

Methods

new()

initialize method.

checkIntegrity()

This method checks whether the identifier used in annotation and data agree and ensures data integrity between the two.

setSubsetAndSortOptions()

set the subset and sortOptions

subsetAnnotation(ids)

given a vector of ids, subset both the data and annotation slots.

setAnnotationDisplayOptions(annotCols)

given a vector of column names in annotation, select those columns and drop the rest.

Examples

See \code{\link{qcFlowObj}}, \code{\link{gatingObj}}, and \code{\link{populationExpressionObj}} for code examples.

#> Error: <text>:1:5: unexpected input
#> 1: See \
#>         ^